NATIVE-325: Add pull to refresh to landing
