+++ src/modules/app/components/Navigator.js
@@ -97,7 +97,7 @@ type PropsType = {|
   setContentLanguage: (language: string) => void,
   fetchCategory: (cityCode: string, language: string, key: string) => void,
   clearCategory: (key: string) => void,
-  fetchCities: () => void
+  fetchCities: (forceRefresh: boolean) => void
 |}
 
 class Navigator extends React.Component<PropsType> {
@@ -113,7 +113,7 @@ class Navigator extends React.Component<PropsType> {
   }
 
   componentDidMount () {
-    this.props.fetchCities()
+    this.props.fetchCities(false)
     this.loadSelectedCity()
   }
